40% faster AVL searching and DICTIONARY operations
